requirements.txt
Flask==2.0.1
Flask-Mail==0.9.1
python-dotenv==0.19.0
gunicorn
tensorflow
Werkzeug==2.2.3
